Use newly implemented UnconfiguredCmvnProject in configure phase.
Better separation of concerts, UnconfiguredCmvnProject's only purpose is to configure projects recursive. We now are able to retrieve the local repository path when only a Maven settings file is given. So, e.g. the Cmvn Eclipse Plugin is able to provide a Classpath Container even when Cmvn was configured with --maven-settings.

Also, we now have a --verbose option to show somewhat more information about what cmvn does.